requirements/django.txt
blinker
django
djangorestframework
django-redis
elasticsearch
gunicorn
mongoengine
psycopg2
python-dateutil
python-social-auth
requests
